+++ src/plugins/security/model/presentation/longdecimaltextbox.h
@@ -18,42 +18,41 @@
  **
  ***********************************************************************************************************************/
 
-#ifndef SECURITY_DATA_ELEMENT_CONTENT_H
-#define SECURITY_DATA_ELEMENT_CONTENT_H
+#ifndef SECURITY_LONG_DECIMAL_TEXTBOX_H
+#define SECURITY_LONG_DECIMAL_TEXTBOX_H
 
-#include "security.h"
+#include "dataelementcontent.h"
 
-#include <string>
+namespace security {
 
-namespace security
-{
-
-/*!
-   \class DataElementContent
- * \brief The DataElementContent class base class for graphical element such as combobox or radio-button.
+/**
+ * \class LongDecimalTextBox
+ * \brief Represents a text box with or without a spin control for entering long decimal numbers.
  *
  * \ingroup security
  * \ingroup presentation
  */
-class DataElementContent
+class LongDecimalTextBox: public DataElementContent
 {
 public:
-    /*!
-     * \brief value An initial value of graphical element.
-     */
-    std::string value;
 
     /*!
-     * \brief refId A reference Id.
+     * \brief defaultValue
      */
-    std::string refId;
+    uint64_t defaultValue;
 
     /*!
-     * \brief postfix A description that follows graphical element.
+     * \brief spinBox
      */
-    std::string postfix;
+    bool spinBox { false };
+
+    /*!
+     * \brief spinStep
+     */
+    uint64_t spinStep {1};
+
 };
 
 } // of namespace security
 
-#endif // SECURITY_DATA_ELEMENT_CONTENT_H
+#endif // SECURITY_LONG_DECIMAL_TEXTBOX_H
