+++ src/plugins/security/model/sdmx/textelement.h
